Add files via upload
The sportsandmore-project-051819 file is the code I have on my end so far. The others are templates I'm trying to work with. Take a look at carrental_final_code -- this is the example from the Packt site. I'm going to try to mess more with this one. I also found the django-locallibrary-tutorial-master and django-reservation examples that could come in handy.